--- 02.typescript/src/pages/update/TodoUpdate.js
@@ -1,1 +1,0 @@
-// 할일 수정
